+++ src/brushes/pencil_brush.class.js
@@ -76,6 +76,7 @@
           }
           this.oldEnd = this._drawSegment(ctx, points[length - 2], points[length - 1], true);
           ctx.stroke();
+          this._drawClipPath(ctx);
           ctx.restore();
         }
       }
@@ -140,13 +141,11 @@
     /**
      * Draw a smooth path on the topCanvas using quadraticCurveTo
      * @private
+     * @param {CanvasRenderingContext2D} ctx
      */
-    _render: function() {
-      var ctx  = this.canvas.contextTop, i, len,
-          p1 = this._points[0],
-          p2 = this._points[1];
+    render: function(ctx) {
+      var i, len, p1 = this._points[0], p2 = this._points[1];
 
-      this._saveAndTransform(ctx);
       ctx.beginPath();
       //if we only have 2 points in the path and they are the same
       //it means that the user only clicked the canvas without moving the mouse
@@ -173,7 +172,6 @@
       // the bezier control point
       ctx.lineTo(p1.x, p1.y);
       ctx.stroke();
-      ctx.restore();
     },
 
     /**
@@ -215,7 +213,7 @@
         this.shadow.affectStroke = true;
         path.shadow = new fabric.Shadow(this.shadow);
       }
-
+      this._addClipPathToResult(path);
       return path;
     },